Bump to latest electron version, ESM conversion
