+++ src/scripts/utils.js
@@ -1,18 +1,14 @@
-const parseGameName = (url) => {
+export const parseGameName = (url) => {
   // Adjusted regex to correctly capture the game name part of the URL
   const match = url.match(/ynoproject\.net\/([^/]+)/);
   return match ? match[1] : null;
 };
 
-const isDreamWorldMap = (url) => {
+export const isDreamWorldMap = (url) => {
     const imageExtensions = [".png", ".jpg", ".jpeg"];
     let isImage = false;
     if (url.startsWith("https://yume.wiki")) {
       isImage = imageExtensions.some((ext) => url.endsWith(ext));
     }
     return isImage;
-};
-
-module.exports = {
-  parseGameName, isDreamWorldMap
 };
